yolo-agent/requirements.txt
gradio[mcp]
ultralytics
opencv-python
numpy
pillow